Revert "x64 fixes, now loads to menu"
This reverts commit 29045da81c1afad13be8e33018f2ad155432870f.

+++ src/gameui/GameConsole.cpp
@@ -141,7 +141,7 @@ void CGameConsole::ActivateDelayed(float time)
 #endif
 }
 
-void CGameConsole::SetParent( intp parent )
+void CGameConsole::SetParent( int parent )
 {	
 #ifndef _XBOX
 	if (!m_bInitialized)
